1.移除pipeline相关的代码 2.重构httpserver/client等相关结构代码  3.添加Module基类

+++ include/utils/FFL_File.hpp
@@ -45,7 +45,7 @@ namespace FFL {
 		//
 		//  文件大小
 		//
-		int32_t getSize();
+		size_t getSize();
 	private:
 		//
 		//  打开文件，FFL_OK成功
